requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
idna==3.7
loguru==0.7.2
oauthlib==3.2.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
urllib3==2.2.1
django-debug-toolbar==4.3.0
django-cors-headers==4.3.1
nanoid==2.0.0